requirements.txt
appdirs==1.4.4
cachelib==0.1.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cs50==6.0.2
distlib==0.3.1
env==0.1.0
ffmpeg-python==0.2.0
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
future==0.18.2
google-api-core==1.27.0
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySimpleGUI==4.34.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
termcolor==1.1.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.2.2
Werkzeug==1.0.1
